server/requirement.txt
Flask==2.2.3
joblib==1.3.1
langchain==0.0.165
cohere==4.1.0
chromadb==0.4.1
pypdf==3.8.0
tiktoken==0.3.3
gtts==2.2.3
tenacity==8.2.2
python-dotenv==1.0.0
